Remove general font-size from Fractal
This is more verbose and slightly more fragile in terms of future changes, but I prefer this approach instead of risking font-size cascading into and affecting the DNA components (like what's happening at the moment). I've looked through all of the documentation and this doesn't seem to cause any issues.